requirements.txt
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
twilio==9.0.0
google-genai==0.2.2
python-dotenv==1.0.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
requests==2.31.0
pytz==2024.1
apscheduler==3.10.4
